Safari-611.1.10/LayoutTests/fast/forms/radio-input-in-shadow-root-crash.html

<!DOCTYPE html>
<html>
<body>
    <script src="../../resources/js-test.js"></script>
    <input type="radio" name="foo" />
    <script>
        description("This test checks that the checked attribute can be set on a disconnected radio button inside a shadow root. This test passes if it does not crash.");

        input = document.querySelector("input")
        const container = document.createElement("div");
        container.attachShadow({ mode: "open" }).appendChild(input);
        input.checked = true;
        shouldBeTrue("input.checked");
    </script>
</body>
</html>
